Feat: Se implemento acceso basado en roles utilizando ASP.NET Core Identity.

+++ TechNotes.Application/Notes/DeleteNote/DeleteNoteCommandHandler.cs
@@ -1,14 +1,19 @@
+using TechNotes.Application.Users;
+
 namespace TechNotes.Application.Notes.DeleteNote;
 
-public class DeleteNoteCommandHandler : ICommandHandler<DeleteNoteCommand>
+public class DeleteNoteCommandHandler(
+    INoteRepository _noteRepository,
+    IUserService _userService
+    ) : ICommandHandler<DeleteNoteCommand>
 {
-    private readonly INoteRepository _noteRepository;
-    public DeleteNoteCommandHandler(INoteRepository noteRepository)
-    {
-        _noteRepository = noteRepository;
-    }
     public async Task<Result> Handle(DeleteNoteCommand request, CancellationToken cancellationToken)
     {
+        var currentUserCanDelete = await _userService.CurrentUserCanEditNoteAsync(request.Id);
+        if (!currentUserCanDelete)
+        {
+            return Result.Fail<NoteResponse?>("No tiene permiso para eliminar esta nota");
+        }
         var deleted = await _noteRepository.DeleteNoteAsync(request.Id);
         if (deleted)
         {
